Refactor token table display
Filter the tokens into their four types and display them as four
separate tables.  Move the token component into its own file and
avoid memoizing the column definition since it's now fixed.
